[Refactor] : Navbar component separated
